+++ src/component/preview/DatasetPreview.svelte
@@ -1,23 +1,23 @@
 <script>
   import db from "@db"
+  import { tab_selected } from "@js/store"
   import Preview_manager from "@js/Preview_manager"
   import Datatable from "@datatable/Datatable.svelte"
   import Loading from "@frame/Loading.svelte"
 
   export let dataset_preview
-  export let nb_item = false
   export let load_first = false
 
   let dataset_data = false
   let columns = []
 
   const load_preview = (async () => {
-    nb_item = 0
+    $tab_selected.nb = 0
     if (typeof dataset_preview !== "string") {
       dataset_data = dataset_preview
       dataset_data = Preview_manager.add_position(dataset_data)
       columns = Preview_manager.get_columns(dataset_data)
-      nb_item = dataset_data.length
+      $tab_selected.nb = dataset_data.length
       return
     }
 
@@ -33,7 +33,7 @@
       return obj
     })
     columns = Preview_manager.get_columns(dataset_data)
-    nb_item = dataset_data.length
+    $tab_selected.nb = dataset_data.length
   })()
 </script>
 
@@ -42,13 +42,12 @@
 {:then}
   {#if dataset_data.length > 0}
     <Datatable
-      entity="dataset_preview"
+      entity="preview"
       data={dataset_data}
       sort_by_name={false}
       {columns}
       keep_all_cols={true}
       {load_first}
-      bind:nb_item
     />
   {/if}
 {/await}