+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - fixed : remove stat tab menu border on dark mode
 - fixed : remove mouse cursor when datatables not clickable
+- fixed : tab nb item update from child component with store
 
 ## 0.8.17 (2024-10-20)